open-webui-60 Remove certain chat-menu-option for 'user' profile from Navbar and Sidebar

+++ src/lib/components/layout/Sidebar/ChatMenu.svelte
@@ -295,7 +295,8 @@
 				</DropdownMenu.Item>
 			{/if}
 
-			<DropdownMenu.Sub>
+			{#if $user?.role === 'admin'}
+				<DropdownMenu.Sub>
 				<DropdownMenu.SubTrigger
 					class="flex gap-2 items-center px-3 py-2 text-sm  cursor-pointer hover:bg-gray-50 dark:hover:bg-gray-800 rounded-md"
 				>
@@ -338,7 +339,9 @@
 					</DropdownMenu.Item>
 				</DropdownMenu.SubContent>
 			</DropdownMenu.Sub>
-			<DropdownMenu.Item
+			{/if}
+			{#if $user?.role === 'admin'}
+				<DropdownMenu.Item
 				class="flex  gap-2  items-center px-3 py-1.5 text-sm  cursor-pointer hover:bg-gray-50 dark:hover:bg-gray-800 rounded-md"
 				on:click={() => {
 					deleteHandler();
@@ -347,6 +350,7 @@
 				<GarbageBin strokeWidth="2" />
 				<div class="flex items-center">{$i18n.t('Delete')}</div>
 			</DropdownMenu.Item>
+			{/if}
 
 			<hr class="border-gray-100 dark:border-gray-850 my-0.5" />
